[ins] prevent propagation if there are no measurement updates
To prevent the position and velocity estimation from running away while you only have accel measurements (done in the propagation),
Propagate only if we got any measurement during the last INS_MAX_PROPAGATION_STEPS.
Otherwise halt the propagation to not diverge and only set the acceleration.
This should only be relevant in the startup phase when the baro is not yet initialized
and there is no gps fix yet...

Should fix the last remaining part of #1237 in a more generic and robust way.