A temporary file for numerical differentiation of simulation data
This file numerically differentiates position and velocity data to check the zero points of the measured velocity and acceleration.
This is to understand the shifts of the reference (desired) and measured values, and why there is no shift between acceleration
curves in the built-in methods of paparazzi. This is a temporary file that can be deleted after the acceleration shift issue is solved.